+++ WMS/Scheduler/Core/Models/Item/ItemWithdrawOptions.cs
@@ -16,11 +16,11 @@ namespace Ferretto.WMS.Scheduler.Core.Models
 
         public int? PackageTypeId { get; set; }
 
-        public int Quantity { get; set; }
+        public double Quantity { get; set; }
 
         public string RegistrationNumber { get; set; }
 
-        public int RequestedQuantity { get; set; }
+        public double RequestedQuantity { get; set; }
 
         public bool RunImmediately { get; set; }
