+++ WMS/Scheduler/Core/Services/SchedulerService.cs
@@ -159,7 +159,7 @@ namespace Ferretto.WMS.Scheduler.Core.Services
             throw new System.NotImplementedException();
         }
 
-        public async Task<IOperationResult<Mission>> CompleteItemMissionAsync(int missionId, int quantity)
+        public async Task<IOperationResult<Mission>> CompleteItemMissionAsync(int missionId, double quantity)
         {
             using (var serviceScope = this.scopeFactory.CreateScope())
             {
